QVAC-21921 tts-cpp: ACE-Step CPU parity fixes (timbre token, seed, sampling)
Follow-up to the initial ACE-Step port: brings the native CPU pipeline to
audio parity with the acestep.cpp CLI.

- cond encoder: load the canonical silence_latent frame from the DiT GGUF and
  feed it as the timbre token for text-to-music (enc_hidden 194 -> 195 tokens,
  matching upstream); detect use_timbre_cls from GGUF metadata
  (acestep.encoder_hidden_size) instead of assuming it off.
- engine: widen GenerateParams::seed to int64 so uint32-range seeds no longer
  overflow (was producing uncorrelated noise); feed the silence frame into the
  cond forward; add env-gated debug dump/inject hooks for parity isolation.
- music-cli: parse seeds with strtoll (int64), expose --temp/--cfg/--topk/--topp
  and --no-phase1 flags, and default steps/shift/duration to 0 (engine
  auto-detect turbo 8/3.0 vs base/sft 50/1.0).

+++ tts-cpp/include/tts-cpp/acestep/engine.h
@@ -69,7 +69,7 @@ struct GenerateParams {
     int         bpm = 0;                 // optional; 0 => N/A (LM/DiT infer)
     std::string keyscale;                // optional, e.g. "C major"
     std::string timesignature;           // optional, e.g. "4/4"
-    int         seed = -1;               // <0 = random
+    long long   seed = -1;               // <0 = random (uint32 range: torch/philox parity)
     // LM sampling (Phase-2 audio codes). Defaults mirror acestep.cpp.
     float       lm_temperature = 0.85f;
     float       lm_top_p       = 0.9f;
